requirements.txt
requests
azure-core
msrest
azure-identity
